refactor(tests): split test_event_dispatch_tick.py into focused modules
Replaces the 1 383-line monolith with five files, each under the
500-line CI limit, and removes it from the GRANDFATHERED_FILES allowlist
in scripts/checks/file_size.py.

Split:
- test_action_executor.py       — TestActionExecutor
- test_event_dispatcher.py      — TestEventDispatcher
- test_tick_scheduler.py        — TestTickScheduler + TestTickSchedulerProductionDefaults
- test_server_persona_wiring.py — TestResolveAgentType + TestAgentServerPersonaLifecycle
                                  + TestInitializePersonaAgents
- test_dispatch_lifecycle.py    — TestCrossAgentMemoryIsolation
                                  + TestEventActionMemoryCycle + TestPerDispatchTimeout

No test logic changed; all helpers (_make_client, _make_agent, persona
configs) are duplicated into each module so each file is self-contained.
