+++ FILEMAP.md
@@ -1,7 +1,7 @@
 # Repository File Map
 
 > Auto-generated by `scripts/generate_filemap.py` — 2026-04-25  
-> **402** tracked files
+> **406** tracked files
 
 ## Summary
 
@@ -10,7 +10,7 @@
 | `internal/` | 93 |
 | `docs/` | 81 |
 | `agents/` | 77 |
-| `tests/` | 61 |
+| `tests/` | 65 |
 | `(root)` | 20 |
 | `scripts/` | 17 |
 | `.github/` | 15 |
@@ -26,7 +26,7 @@
 
 | Extension | Count |
 |-----------|-------|
-| `.py` | 149 |
+| `.py` | 153 |
 | `.md` | 97 |
 | `.go` | 90 |
 | `.yaml` | 25 |
@@ -479,6 +479,7 @@
 │           ├── __init__.py
 │           ├── _persona_test_helpers.py
 │           ├── conftest.py
+│           ├── test_action_executor.py
 │           ├── test_agents.py
 │           ├── test_base_agent.py
 │           ├── test_base_handle.py
@@ -486,6 +487,7 @@
 │           ├── test_chat_servicer.py
 │           ├── test_defaults.py
 │           ├── test_dispatch_execute_actions.py
+│           ├── test_dispatch_lifecycle.py
 │           ├── test_docker_fixes.py
 │           ├── test_episodic_memory_compression.py
 │           ├── test_episodic_memory_core.py
@@ -493,7 +495,7 @@
 │           ├── test_episodic_memory_min_score.py
 │           ├── test_episodic_memory_retention.py
 │           ├── test_episodic_memory_validation.py
-│           ├── test_event_dispatch_tick.py
+│           ├── test_event_dispatcher.py
 │           ├── test_llm_client.py
 │           ├── test_llm_persona_agent.py
 │           ├── test_memory_injection.py
@@ -514,7 +516,9 @@
 │           ├── test_relationship_memory_user.py
 │           ├── test_sandbox.py
 │           ├── test_server.py
+│           ├── test_server_persona_wiring.py
 │           ├── test_task_agent.py
+│           ├── test_tick_scheduler.py
 │           ├── test_tools.py
 │           ├── test_validate.py
 │           └── test_working_memory.py
